Split dummyData into separate files and moved it into the modules, since having it at a higher level resulted in circular dependencies.
