+++ app/modules/content-items/dummyData.js
@@ -1,29 +1,10 @@
 // @flow
 
-import type { TopicsState } from 'modules/topics';
+import { contentItemTypes, highlightTypes } from './model';
 import type {
-  ContentItemsState,
   ParagraphContentItem,
-} from 'modules/content-items';
-import {
-  contentItemTypes,
-  highlightTypes,
-} from 'modules/content-items';
-
-export const dummyTopics: TopicsState = {
-  feeyhnd0w0: {
-    id: 'feeyhnd0w0',
-    title: 'Test topic 1',
-    description: 'Lorem ipsum dolor sit amet.',
-    rootContentItemId: 'w4lg2u0p1h',
-  },
-  exrhrl5gvy: {
-    id: 'exrhrl5gvy',
-    title: 'Test topic 2',
-    description: '',
-    rootContentItemId: 'qyrgv0bcd6',
-  },
-};
+  ContentItemsState,
+} from './model';
 
 export const dummyContentItems: ContentItemsState = {
   w4lg2u0p1h: {
