+++ app/modules/topics/dummyData.js
@@ -1,0 +1,18 @@
+// @flow
+
+import type { TopicsState } from './model';
+
+export const dummyTopics: TopicsState = {
+  feeyhnd0w0: {
+    id: 'feeyhnd0w0',
+    title: 'Test topic 1',
+    description: 'Lorem ipsum dolor sit amet.',
+    rootContentItemId: 'w4lg2u0p1h',
+  },
+  exrhrl5gvy: {
+    id: 'exrhrl5gvy',
+    title: 'Test topic 2',
+    description: '',
+    rootContentItemId: 'qyrgv0bcd6',
+  },
+};
